Create Radha, has bought 5 books for reference to prepare for her semester exams. She has studied in mathematics course that the factorial of n items gives the number of ways that she can use to arrange the n items. Radha, would like to know in how many ways she can arrange the five books that she has bought Following are requirements to solve the problem using while loop: a.	Number of books has to be captured b.	The books should be arranged starting from the first book and then the second book so on c.	Finally, the total number of ways in which her books can be arranged should be  displayed.
Radha, has bought 5 books for reference to prepare for her semester exams. She has studied in mathematics course that the factorial of n items gives the number of ways that she can use to arrange the n items. Radha, would like to know in how many ways she can arrange the five books that she has bought
Following are requirements to solve the problem using while loop:
a.	Number of books has to be captured
b.	The books should be arranged starting from the first book and then the second book so on
c.	Finally, the total number of ways in which her books can be arranged should be  displayed.